requirements.txt
cachetools==5.2.0
celery==5.2.7
django==3.2.16
django-authtools==2.0.0
django-celery-beat==2.3.0
django-cors-headers==3.13.0
django-debug-toolbar
django-debug-toolbar-force
django-environ==0.9.0
django-filter==22.1
django-model-utils==4.2.0
django-redis==5.2.0
django-timezone-field==5.0
djangorestframework-camel-case==1.3.0
djangorestframework==3.14.0
docutils==0.19
drf-yasg[validation]==1.21.4
gunicorn[gevent]==20.1.0
hexbytes==0.2.2
lxml==4.9.1
numpy==1.23.3
packaging==21.3
psycogreen==1.0.2
psycopg2-binary==2.9.1
psycopg2==2.9.4
redis==4.3.4
requests==2.28.1
sgqlc==11.0
safe-eth-py[django]==4.5.1
web3==5.31.1